Move search code into its own module
