+++ intake_esm/core.py
@@ -1,17 +1,15 @@
-import itertools
 import json
 import logging
-from collections.abc import Iterable
-from typing import Any, Dict, List, Pattern, Union
+from typing import Any, Dict, List, Union
 from warnings import warn
 
 import dask
 import intake
-import numpy as np
 import pandas as pd
 import xarray as xr
 from fastprogress.fastprogress import progress_bar
 
+from .search import _unique, search
 from .utils import _fetch_and_parse_json, _fetch_catalog, logger
 
 
@@ -402,7 +400,7 @@ class esm_datastore(intake.catalog.Catalog):
         4    landCoverFrac
         """
 
-        results = _get_subset(self.df, require_all_on=require_all_on, **query)
+        results = search(self.df, require_all_on=require_all_on, **query)
         ret = esm_datastore.from_df(
             results,
             esmcol_data=self.esmcol_data,
@@ -668,107 +666,6 @@ class esm_datastore(intake.catalog.Catalog):
                 progress.update(total)
 
             return self._datasets
-
-
-def _unique(df, columns=None):
-    if isinstance(columns, str):
-        columns = [columns]
-    if not columns:
-        columns = df.columns.tolist()
-    info = {}
-    for col in columns:
-        values = df[col].dropna().values
-        uniques = np.unique(list(_flatten_list(values))).tolist()
-        info[col] = {'count': len(uniques), 'values': uniques}
-    return info
-
-
-def _get_subset(df, require_all_on=None, **query):
-    message = 'Query returned zero results.'
-    if not query:
-        warn(message)
-        return pd.DataFrame(columns=df.columns)
-    condition = np.ones(len(df), dtype=bool)
-    query = _normalize_query(query)
-    for key, val in query.items():
-        condition_i = np.zeros(len(df), dtype=bool)
-        column_is_stringtype = isinstance(
-            df[key].dtype, (np.object, pd.core.arrays.string_.StringDtype)
-        )
-        for val_i in val:
-            value_is_pattern = _is_pattern(val_i)
-            if column_is_stringtype and value_is_pattern:
-                cond = df[key].str.contains(val_i, regex=True, case=True, flags=0)
-            else:
-                cond = df[key] == val_i
-            condition_i |= cond
-        condition &= condition_i
-    query_results = df.loc[condition]
-
-    if require_all_on:
-        if isinstance(require_all_on, str):
-            require_all_on = [require_all_on]
-        _query = query.copy()
-
-        # Make sure to remove columns that were already
-        # specified in the query when specified in `require_all_on`. For example,
-        # if query = dict(variable_id=["A", "B"], source_id=["FOO", "BAR"])
-        # and require_all_on = ["source_id"], we need to make sure `source_id` key is
-        # not present in _query for the logic below to work
-        for key in require_all_on:
-            _query.pop(key, None)
-
-        keys = list(_query.keys())
-        grouped = query_results.groupby(require_all_on)
-        values = [tuple(v) for v in _query.values()]
-        condition = set(itertools.product(*values))
-        results = []
-        for key, group in grouped:
-            index = group.set_index(keys).index
-            if not isinstance(index, pd.MultiIndex):
-                index = {(element,) for element in index.to_list()}
-            else:
-                index = set(index.to_list())
-            if index == condition:
-                results.append(group)
-
-        if len(results) >= 1:
-            return pd.concat(results).reset_index(drop=True)
-
-        warn(message)
-        return pd.DataFrame(columns=df.columns)
-    if query_results.empty:
-        warn(message)
-
-    return query_results.reset_index(drop=True)
-
-
-def _normalize_query(query):
-    q = query.copy()
-    for key, val in q.items():
-        if isinstance(val, str) or not isinstance(val, Iterable):
-            q[key] = [val]
-    return q
-
-
-def _is_pattern(value):
-    value_is_repattern = isinstance(value, Pattern)
-    if value_is_repattern:
-        return True
-    wildcard_chars = {'*', '?', '$', '^'}
-    try:
-        return any(char in value for char in wildcard_chars)
-    except TypeError:
-        return False
-
-
-def _flatten_list(data):
-    for item in data:
-        if isinstance(item, Iterable) and not isinstance(item, str):
-            for x in _flatten_list(item):
-                yield x
-        else:
-            yield item
 
 
 def _make_entry(key, df, aggregation_info):
